fix: download progress integer overflow + terminal black screen
- Fix int overflow: totalRead * 100 overflowed int at ~20.5MB,
  cast to long for correct percentage calculation
- Fix black screen: TermuxSplitLayout was missing onMeasure() and
  had unconditional black background fill in onDraw() that painted
  over all TerminalView children
- Remove mBackgroundPaint and BACKGROUND_COLOR from SplitLayout
- Add proper onMeasure() that measures children with EXACTLY spec

+++ app/src/main/java/com/termux/app/TermuxActivity.java
@@ -7477,7 +7477,7 @@ public final class TermuxActivity extends AppCompatActivity implements ServiceCo
                                 sizeText.setText(readMb + " MB received");
                             });
                         } else {
-                            final int pct = Math.min(100, totalRead * 100 / Math.max(1, totalSize));
+                            final int pct = (int) Math.min(100, (long) totalRead * 100 / Math.max(1L, (long) totalSize));
                             final int readMb = totalRead / (1024 * 1024);
                             final int totalMb = totalSize / (1024 * 1024);
                             runOnUiThread(() -> {
